requirements.txt
google-generativeai==0.3.0
python-dotenv==1.0.0
flask==2.3.2
werkzeug==2.3.6
